requirements.txt
astroid==2.3.3
colorzero==1.1
gpiozero==1.5.1
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pythonping==1.0.9
schedule==0.6.0
six==1.14.0
speedtest-cli==2.1.2
typed-ast==1.4.1
wrapt==1.11.2
